+++ framework/i18n/src/test/java/is/codion/framework/i18n/FrameworkMessagesTest.java
@@ -16,7 +16,6 @@ public class FrameworkMessagesTest {
     FrameworkMessages.exitTip();
     FrameworkMessages.view();
     FrameworkMessages.viewMnemonic();
-    FrameworkMessages.refreshAll();
     FrameworkMessages.supportTables();
     FrameworkMessages.supportTablesMnemonic();
     FrameworkMessages.update();